fix printf errors / warnings for bare-metal

+++ apps/demos/src/atc.c
@@ -4,15 +4,15 @@
 #include "atc.h"
 
 void read_atc_registers(){
-  printf("timing status---------------0x%x \r\n", axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_STATUS));
-  printf("timestamp-------------------0x%x \r\n", axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_TIMESTAMP));
-  printf("polarity--------------------0x%x \r\n", axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_POLARITY));
-  printf("destination LEMO A----------0x%x \r\n", axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_DST_LEMO_A));
-  printf("destination LEMO B----------0x%x \r\n", axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_DST_LEMO_B));
-  printf("destination poke C----------0x%x \r\n", axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_DST_POKE_C));
-  printf("destination poke D----------0x%x \r\n", axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_DST_POKE_D));
-  printf("destination logic E---------0x%x \r\n", axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_DST_LOGIC_E));
-  printf("destination logic F---------0x%x \r\n", axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_DST_LOGIC_F));
+  printf("timing status---------------0x%x \r\n", (unsigned int) axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_STATUS));
+  printf("timestamp-------------------0x%x \r\n", (unsigned int) axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_TIMESTAMP));
+  printf("polarity--------------------0x%x \r\n", (unsigned int) axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_POLARITY));
+  printf("destination LEMO A----------0x%x \r\n", (unsigned int) axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_DST_LEMO_A));
+  printf("destination LEMO B----------0x%x \r\n", (unsigned int) axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_DST_LEMO_B));
+  printf("destination poke C----------0x%x \r\n", (unsigned int) axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_DST_POKE_C));
+  printf("destination poke D----------0x%x \r\n", (unsigned int) axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_DST_POKE_D));
+  printf("destination logic E---------0x%x \r\n", (unsigned int) axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_DST_LOGIC_E));
+  printf("destination logic F---------0x%x \r\n", (unsigned int) axil_read_register(C_SCOPE_ATC + C_ADDR_ATC_DST_LOGIC_F));
 }
 
 #define C_ATC_BUSY_WAIT 10
@@ -29,35 +29,35 @@ void read_atc_counts(){
   axil_write_register(C_SCOPE_ATC+C_ADDR_ATC_COUNT_REQ, 0b01110000);
   wait_atc_busy(C_ATC_BUSY_WAIT);
   count = axil_read_register(C_SCOPE_ATC+C_ADDR_ATC_COUNT);
-  printf("LEMO A----------------------%4d (0x%x) \r\n", count, count);
+  printf("LEMO A----------------------%4d (0x%x) \r\n", count, (unsigned int) count);
 
   axil_write_register(C_SCOPE_ATC+C_ADDR_ATC_COUNT_REQ, 0b01110001);
   wait_atc_busy(C_ATC_BUSY_WAIT);
   count = axil_read_register(C_SCOPE_ATC+C_ADDR_ATC_COUNT);
-  printf("LEMO B----------------------%4d (0x%x) \r\n", count, count);
+  printf("LEMO B----------------------%4d (0x%x) \r\n", count, (unsigned int) count);
 
   axil_write_register(C_SCOPE_ATC+C_ADDR_ATC_COUNT_REQ, 0b01110010);
   wait_atc_busy(C_ATC_BUSY_WAIT);
   count = axil_read_register(C_SCOPE_ATC+C_ADDR_ATC_COUNT);
-  printf("POKE C----------------------%4d (0x%x) \r\n", count, count);
+  printf("POKE C----------------------%4d (0x%x) \r\n", count, (unsigned int) count);
 
   axil_write_register(C_SCOPE_ATC+C_ADDR_ATC_COUNT_REQ, 0b01110011);
   wait_atc_busy(C_ATC_BUSY_WAIT);
   count = axil_read_register(C_SCOPE_ATC+C_ADDR_ATC_COUNT);
-  printf("POKE D----------------------%4d (0x%x) \r\n", count, count);
+  printf("POKE D----------------------%4d (0x%x) \r\n", count, (unsigned int) count);
 
   for (int i=0; i<10; i++){
     axil_write_register(C_SCOPE_ATC+C_ADDR_ATC_COUNT_REQ, 0b01010000 + i);
     wait_atc_busy(C_ATC_BUSY_WAIT);
     count = axil_read_register(C_SCOPE_ATC+C_ADDR_ATC_COUNT);
-    printf("OUTPUT G(%d)-----------------%4d (0x%x) \r\n", i, count, count);
+    printf("OUTPUT G(%d)-----------------%4d (0x%x) \r\n", i, count, (unsigned int) count);
   }
 
   for (int i=0; i<10; i++){
     axil_write_register(C_SCOPE_ATC+C_ADDR_ATC_COUNT_REQ, 0b01100000 + i);
     wait_atc_busy(C_ATC_BUSY_WAIT);
     count = axil_read_register(C_SCOPE_ATC+C_ADDR_ATC_COUNT);
-    printf("OUTPUT H(%d)-----------------%4d (0x%x) \r\n", i, count, count);
+    printf("OUTPUT H(%d)-----------------%4d (0x%x) \r\n", i, count, (unsigned int) count);
   }
 }
